Rename cryo plant path
